move cli package
