+++ config/makefile/cli.mk
@@ -1,0 +1,4 @@
+.PHONY: run-cli
+
+run-cli:
+	cd metal/cli && go run main.go
