+++ metal/cli/main.go
@@ -2,9 +2,9 @@ package main
 
 import (
 	"fmt"
-	"github.com/oullin/cli/accounts"
-	"github.com/oullin/cli/panel"
-	"github.com/oullin/cli/posts"
+	"github.com/oullin/metal/cli/accounts"
+	"github.com/oullin/metal/cli/panel"
+	"github.com/oullin/metal/cli/posts"
 	"github.com/oullin/database"
 	"github.com/oullin/env"
 	"github.com/oullin/metal/kernel"
@@ -17,7 +17,7 @@ var environment *env.Environment
 var dbConn *database.Connection
 
 func init() {
-	secrets := kernel.Ignite("./../.env", pkg.GetDefaultValidator())
+	secrets := kernel.Ignite("./../../.env", pkg.GetDefaultValidator())
 
 	environment = secrets
 	dbConn = kernel.MakeDbConnection(environment)
