+++ metal/cli/panel/menu_test.go
