+++ metal/cli/posts/factory.go
