+++ metal/cli/posts/handler_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/oullin/cli/clitest"
+	"github.com/oullin/metal/cli/clitest"
 	"github.com/oullin/database"
 	"github.com/oullin/pkg"
 	"github.com/oullin/pkg/markdown"
