+++ metal/cli/posts/input.go
